fix 401 error setting quotas on radosgw swift
In the case of radosgw swift, projects are lazily created until the
first member of the project attempts to access the object store. This
updates the code to handle 401 unauthorized errors by creating a
temporary app credential in the project used to connect to the object
store to populate the account. From there the code can proceed as usual.