Warn but don't fail on swift quotas
We still haven't fully ironed out the implementation of
quotas for RGW based object storage. But in that scenario, applying
quotas fails and the executions stops and fails to apply volume
quotas.

We want to warn that something is wrong, but not outright fail,
so that execution can continue and volume quotas apply.