Include Coldfront project name when generating project display name
Add function to sanitize project name. Move `get_unique_project_name` to
utils.py, and then use these 2 functions to get generate acceptable and unique
project names for openshift projects.

Also redo the return type of create_project so that it returns a namedtuple
with project_name and project_id, this allows the different resources to
modify the project name.

closes https://github.com/nerc-project/coldfront-plugin-cloud/issues/87