Remove ubuntu-specific commands from ci/microshift.sh
By making ci/microshift.sh only about deploying Microshift, we make it more
useful to developers who may be running a non-Ubuntu Linux distribution
locally. This allows them to run the CI tests without also needing to spin
up an Ubuntu virtual machine.

Note that with this commit we're assuming that Docker is already
available, which is true for the Github actions environment. The
assumption is that developers running the code locally will provide an
environment with Docker.