Add authentication for REST API
This limits the Allocations REST API to admin users.

In the presence of the PLUGIN_OIDC flag, which signals that ColdFront
has been configured with OpenID Connect, this only allows
authentication via access tokens.

If the flag has not been detected, than it falls back to
Basic and Session authentications.