Update the `validate_allocations` command to sync users in addition to quotas.
To accomplish this, we added a method to retrieve a list of users from
allocators. When a user is missing from the allocator but exists in coldfront
it is correctly added to the allocator. Conversely, when a user that is not
in coldfront is part of an allocation, I have decided that it is best to delete
that user as we expect coldfront to be the source of truth.