fix radosgw CI/CD pipeline failures
This fixes an issue with the radosgw portion of the CI/CD pipeline due
to an upstream issue/change with ceph-exporter related to the metrics
setup on the cluster. This patch simply disables setting up the
monitoring stack altogether given that the metrics setup is not used by
this pipeline.

This patch also doubles the size of the OSDs given that the most recent
version of ceph deployed by cephadm 17.2.6-0ubuntu0.22.04.1 requires
more space. This was determined by the appearance of out-of-space errors
in OSD logs on an external test VM.

In addition to these changes, this patch also disables configuring
firewalld given that it's not installed and not necessary for the CI/CD
pipeline.