Support excluding time intervals during storage invoicing
Added an optional CLI argument which allows passing in a txt file
containing excluded time intervals. Each line in the txt file should
contain a date range that we want to exclude from billing, formatted as
 "%Y-%m-%d,%Y-%m-%d". An example line could be "2023-02-01,2023-02-03".

The time period excluded from billing starts from the first second of
the start date, and ends on the first second of the end date.