Fixed bug for billing with change requests
Due to an boolean statement, the SU hours for storage will not
be correctly calculated when change requests (CRs) to quotas are involved.

Our intention with CRs is that when an allocation's storage quota
decreases, this decreased value becomes effective (for billing purposes)
when the CR for it is created, not when the CR is
approved (or the time at which the quota actually decreased).
The aforementioned boolean statement would instead lead to undesired behavior.