Fixed integration with Openshift API
It was discovered that Openshift integration did not
function because the Openshift allocation would use
the same url to make calls to both the account
manager and the Openshift API. This resulted in the
Openshift API never actually being called. Due to
poor error handling by the integration code (more details below),
this bug went undetected by the functional tests.
Appropriate fixes have been added to tests, and the
Openshift resource type now requires two URLs, one
for the account manager, and one for the Openshift API.

Regarding the poor error handling, none of the functional
Openshift test cases actually checked if the call to
`get_federated_user()` returned the expected output.
The `get_federated_user()` function itself only emits a
log message if the user is not found. This meant that
even though `get_federated_user()` never called the
Openshift API and would therefore never find the user, the test cases still passed.

Additionally, while `_openshift_user_exists()`, the function
which was supposed to call the Openshift API, does
catch the `kexc.NotFoundError`, this error is not specific
enough, as it could be caused by a 404 response made
by ANY server (in the case of our tests, the account manager).

It was also found that the `RESOURCE_IDENTITY_NAME`
attribute, which identifies the idp, was referenced in
integration code but never defined, leading to
`_openshift_identity_exists` always failing, since
`self.id_provider` would always be `None`