Allow pre-existing OpenShift projects to be managed by Coldfront allocations
To use this feature, first create a Coldfront allocation with the
appropriate OpenShift resource. Before approving it, use the admin
dashboard to add the attributes `Allocated Project ID` and `Allocated Project Name`
to the allocation, both assigned to the name of the pre-existing project.
Afterwards, approve the allocation, then run `validate_allocations` to
populate the quota fields (as well as create various other OpenShift objects)
on the allocation and the OpenShift project

Do note, because of how validation is handled for quotas [1], if the
pre-existing project already defined its quotas, Coldfront won't
overwrite them with the default quotas.

[1] https://github.com/nerc-project/coldfront-plugin-cloud/blob/1b2338d309c7db1c5a418d25fc40169b7edae055/src/coldfront_plugin_cloud/management/commands/validate_allocations.py#L311-L316