Management command to fetch daily service invoices
This PR introduces a new management command
`coldfront fetch_daily_billable_usage --date <YYYY-MM>`.

This management command fetches the daily billable usage reports
stored in BackBlaze S3 under the following location
/Invoices/<YYYY-MM>/Service Invoices/<CLUSTER> <YYYY-MM-DD>.csv

For each allocation in either Active or Active (Needs Renewal)
state of the 3 production clusters (NERC, NERC-OCP and NERC-EDU),
the attribute named "Cumulative Daily Charges for Month" is updated
with the latest total billable cost for that allocation for that given
month.

The attribute stores the total in the following format:
<YYYY-MM-DD>: <TOTAL> USD

If an allocation has the attribute Monthly Allocation Alert set to a
value different than zero, an email is going to be sent to the PI ONCE
after the total billable cost for that allocation has been exceeded.